+++ src/extension.ts
@@ -1,12 +1,12 @@
 import * as vscode from 'vscode';
-import * as path from 'path';
-import { Provider, Item, BenchmarkItem, AllocationItem } from './provider';
+import { TreeDataProvider, Item, BenchmarkItem, AllocationItem } from './treedata';
+import { CodeLensProvider } from './codelens';
 
 export async function activate(context: vscode.ExtensionContext) {
-    const provider = new Provider();
+    const treeData = new TreeDataProvider();
 
     let options: vscode.TreeViewOptions<Item> = {
-        treeDataProvider: provider,
+        treeDataProvider: treeData,
         showCollapseAll: true
     }
     let treeView = vscode.window.createTreeView<Item>('goAllocationsExplorer', options);
@@ -35,9 +35,9 @@ export async function activate(context: vscode.ExtensionContext) {
     const runAllBenchmarksSimple = vscode.commands.registerCommand('goAllocations.runAllBenchmarksSimple',
         async () => {
             try {
-                await provider.runAllBenchmarksSimple(treeView);
+                await treeData.runAllBenchmarksSimple(treeView);
             } catch (error) {
-                if (provider.abortSignal().aborted) {
+                if (treeData.abortSignal().aborted) {
                     vscode.window.showInformationMessage('Operation(s) cancelled');
                 } else {
                     console.error('Error running all benchmarks:', error);
@@ -50,16 +50,16 @@ export async function activate(context: vscode.ExtensionContext) {
     const stopAllBenchmarks = vscode.commands.registerCommand('goAllocations.stopAllBenchmarks',
         () => {
             vscode.window.showInformationMessage('Cancelling operation(s). Go processes may take a moment to terminate.');
-            provider.cancelAll();
+            treeData.cancelAll();
         });
     context.subscriptions.push(stopAllBenchmarks);
 
     const runSingleBenchmark = vscode.commands.registerCommand('goAllocations.runSingleBenchmark',
         async (benchmarkItem: BenchmarkItem) => {
-            const signal = provider.abortSignal();
+            const signal = treeData.abortSignal();
 
             try {
-                provider.clearBenchmarkRunState(benchmarkItem);
+                treeData.clearBenchmarkRunState(benchmarkItem);
                 await treeView.reveal(benchmarkItem, { expand: true });
             } catch (error) {
                 if (signal.aborted) {
@@ -75,14 +75,11 @@ export async function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(runSingleBenchmark);
 
     const refresh = vscode.commands.registerCommand('goAllocations.refresh',
-        () => provider.refresh());
+        () => treeData.refresh());
     context.subscriptions.push(refresh);
 
-    // Register a CodeLens provider for Go benchmark functions
-    // Note: CodeLens ordering from multiple providers is determined by the range position.
-    // To appear after the Go extension's lenses, we use a range that starts slightly to the right.
-    const codeLensProvider = vscode.languages.registerCodeLensProvider({ language: 'go', scheme: 'file' }, new GoBenchmarkCodeLensProvider());
-    context.subscriptions.push(codeLensProvider);
+    const codeLens = vscode.languages.registerCodeLensProvider({ language: 'go', scheme: 'file' }, new CodeLensProvider());
+    context.subscriptions.push(codeLens);
 
     // Command invoked by CodeLens in editor to run a specific benchmark
     const runBenchmarkFromEditor = vscode.commands.registerCommand('goAllocations.runBenchmarkFromEditor', async (args: { packageDir: string; benchmarkName: string }) => {
@@ -93,8 +90,8 @@ export async function activate(context: vscode.ExtensionContext) {
             // Focus the Go Allocations Explorer view
             await vscode.commands.executeCommand('workbench.view.extension.goAllocations');
 
-            const benchmarkItem = await provider.findBenchmark(args.packageDir, args.benchmarkName);
-            provider.clearBenchmarkRunState(benchmarkItem);
+            const benchmarkItem = await treeData.findBenchmark(args.packageDir, args.benchmarkName);
+            treeData.clearBenchmarkRunState(benchmarkItem);
             await treeView.reveal(benchmarkItem, { expand: true, select: true });
         } catch (err) {
             const msg = err instanceof Error ? err.message : String(err);
@@ -105,37 +102,3 @@ export async function activate(context: vscode.ExtensionContext) {
 }
 
 export function deactivate() { }
-
-class GoBenchmarkCodeLensProvider implements vscode.CodeLensProvider {
-    private readonly benchRegex = /^\s*func\s+(Benchmark[A-Za-z0-9_]+)\s*\(b\s*\*testing\.B\)/;
-    private onDidChangeCodeLensesEmitter = new vscode.EventEmitter<void>();
-    public readonly onDidChangeCodeLenses: vscode.Event<void> = this.onDidChangeCodeLensesEmitter.event;
-
-    constructor() { }
-
-    provideCodeLenses(document: vscode.TextDocument, _token: vscode.CancellationToken): vscode.ProviderResult<vscode.CodeLens[]> {
-        // Only add to _test.go files
-        if (!document.fileName.endsWith('_test.go')) return [];
-
-        const lenses: vscode.CodeLens[] = [];
-        for (let i = 0; i < document.lineCount; i++) {
-            const line = document.lineAt(i);
-            const m = line.text.match(this.benchRegex);
-            if (m) {
-                const benchName = m[1];
-                // Unfortunately, there is no official API to control CodeLens ordering across providers.
-                // The best we can do is use the same range as the Go extension and hope VS Code's merge
-                // algorithm places ours in a reasonable position.
-                const range = new vscode.Range(i, 0, i, line.text.length);
-                const packageDir = path.dirname(document.uri.fsPath);
-                const cmd: vscode.Command = {
-                    command: 'goAllocations.runBenchmarkFromEditor',
-                    title: 'find allocations',
-                    arguments: [{ packageDir, benchmarkName: benchName }]
-                };
-                lenses.push(new vscode.CodeLens(range, cmd));
-            }
-        }
-        return lenses;
-    }
-}
